changed buffer size to 1024

+++ src/afl-fuzz.c
@@ -2319,7 +2319,7 @@ int main(int argc, char **argv_orig, char **envp) {
       }
       printf("4");
 
-      memcpy(recieved_array, recieve_data.data_buff, BUFF_SIZE_RECEIVER * sizeof(double));
+      memcpy(recieved_array, recieve_data.data_buff, BUFF_SIZE_RECEIVER * sizeof(u32));
       afl->current_entry = recieved_array[0];
       afl->queue_cur = afl->top_rated[afl->current_entry];
       afl->current_entry = afl->queue_cur->id;
